Add interface for providing graph properties guarantees
The idea is that a specialized graph wrappers (such as `Tree`,
`LoopFree`, etc.) will override the default implementation of the
guarantees, while ensuring that such property holds for the graph. Then,
problems can exploit this additional information for using a specialized
algorithm that works only a subcategory of graphs.

It is important to keep in minds that these guarantees are static. That
is, the type must ensure that they hold at all times. On the other hand,
if a type can't guarantee a property, it returns false regardless of the
actual graph structure. That is, it returns false even if the instance
may satisfy the property. This fact guarantees that asking questions
about properties is extremely cheap.

Any code is free to implement runtime checks of these properties if that
is desirable.