Allow custom index type in weak traits
This is a next step for good support of implicit graphs.

It is not complete without allowing custom indices in Neighbors trait
too. However, I don't want to clutter that trait with custom index types
as any support for implicit graph should not limit or complicate
interfaces for normal graphs (the most common use case). IndexType
provides additional properties such as Ord, Eq, etc. which can be relied
on. I will leave this problem unsolved for now.