Do not parametrize is_connected guarantee by edge type
It is intuitive and practical to answer the is_connected property in
regards of the concrete edge of the graph and not let user to ask for
this with arbitrary edge type. It may be useful to ask is connected in
undirected sense to a directed graph, but this may be unnecessarilly
difficult to determine generally and so it is more suitable for an
algorithm rather than a static guarantee.