Fix default implementations for clear and clear_edges
Before this patch, the default implementation could now work correctly
for graphs that do not guarantee stable indices. The new implementation
assumes that the indices remain stable if removing entities in
descending order (this is the same assumption used in Stable::apply) and
that the {vertex,edge}_indices iterators return indices in ascending
order. It is let for future if these assumptions should be codified into
traits or being part of the contract when implementing respective
traits.