Introduce *Base traits
A significant part of Vertices and Edges traits was not dependent on the
corresponding data (types V and E). Extracting them to a separate
supertrait that does not depend on the generic type should simplify many
of the type bounds in all the code, where data-dependent functionality
is not needed and thus present dependencies on either V or E are just to
be able to depend on Vertices<V> and Edges<E>. This is no longer
necessary. Future commits will slowly apply this simplification to all
parts of code where it helps.