Add optional generic type to operators
This is useful for placing there generic types that are necessary for
constraints for implementation of the operator, where the compiler would
complain about the type being unconstrained by the impl trait, self
type, or predicates. It simplifies the type signature of the operators
themselves and there is no need for using PhantomData.