Implement Graph and Path types
Both types are supposed to be user-facing types for working with graphs,
so the users do not use storages directly. `Graph` represents a generic
graph, `Path` represents path-only graphs. The main advantage of using
`Graph` type instead of the storage directly is that it has associated
functions directly on itself, so they are visibly listed in the
documentation. Both types also provides additional convenience
functions, such as new_{un,}directed or stabilize/freeze.