Use Borrow trait to make high-level interface more ergonomic
This allows use of owned indices with high-level API for graphs from
`graph` module, if the index types are Copy, which they will be in the
vast maority of cases. Using owned indices is kind of a convention in
Rust's graph ecosystem.

This technique is not employed in the base traits as that would
complicate the overall interface and result in more code bloat during
monomorphization.