Refactor the approach to weights
The tricky thing is to find compromise between floats not implementing
Ord and algorithms that need Ord weights. We (probably) must have a
special trait anyway (because we want `is_unsigned() -> bool`), which
unfortunately means that not all PartialOrd + Add types can be used as
weights (let's see how much trouble this is in practice).

Nevertheless, requiring only PartialOrd instead of Ord and having an
associated Ord type (that will be Self in majority of cases, except
floats) which can be easily converted from and to the actual type seems
like a reasonable compromise.