Fix and unify undirected self-loop behavior in all storages
As of now, the storages do not yield an undireceted self-loop twice.
This is however an implementation-specific detail and is *not* required
for storages in general. In fact, the default implementation of
Neighbors::degree{,_directed} relies on the
Neighbors::neighbors{,_directed} to tield it twice in order to retrun
correct degree. This should be stated in the documentation.