Allow integer conversions and sentinel value for all id types
This will allow more flexibility of code that uses an id type and can
handle both integer-like and not-like-integer types with significant
performance benefit if the id type is an integer. After this change,
this can be detected at runtime, choosing the appropriate implementation
depending on the actual instance of the graph passed.

And example will be CompactIdMap. For integer indices it can use the
current implementation which has minimal overhead. After this change, it
can support also non-like-integer ids, keeping the mapping from the id
to usize in a hashmap.