Hide non-essential code
I am still not sure about the API of these and I consider them
non-essential to a graph library. I want to focus on the essential
parts, make their API good and write good documentation.

On the other hand, I don't want to delete the code to be able to build
upon it eventually.